+++ ext/CatalystBifurcationKitExtension/bifurcation_kit_extension.jl
@@ -3,7 +3,7 @@
 # Creates a BifurcationProblem, using a ReactionSystem as an input.
 function BK.BifurcationProblem(rs::ReactionSystem, u0_bif, ps, bif_par, args...;
                                plot_var=nothing, record_from_solution=BK.record_sol_default, jac=true, u0=[], kwargs...)
-    if !is_autonomous(rs) 
+    if !isautonomous(rs) 
         error("Attempting to create a `BifurcationProblem` for a non-autonomous system (e.g. where some rate depend on $(rs.iv)). This is not possible.")
     end